requirements.txt
tensorflow
pandas
numpy
scikit-learn
flask
pillow
